Added additional test data to data.sql, created simple validation for request in controller, Get operations now retrieving data correctly.

+++ Backend/src/main/java/com/wcf/entity/Client.java
@@ -14,15 +14,15 @@ public class Client implements Serializable {
 
 	@Id
 	@GeneratedValue
-	private Long clientID;
+	private int clientID;
 	
 	@Column(name="clientname")
 	private String clientName;
 	
-	public Long getClientID() {
+	public int getClientID() {
 		return clientID;
 	}
-	public void setClientID(Long clientID) {
+	public void setClientID(int clientID) {
 		this.clientID = clientID;
 	}
 	public String getClientName() {
